# WaypointQuadtree::final_report