([Anakin & Padme meme, large image](https://www.reddit.com/r/ProgrammerHumor/comments/pq0ia7/oh_the_leaks/))
Part of [making siteupdate "Valgrind clean"](https://github.com/TravelMapping/DataProcessing/issues/375#issuecomment-874435287) is explicitly [`delete`ing dynamically allocated memory. Compared to just letting the OS magically delete everything allocated](https://github.com/TravelMapping/DataProcessing/issues/375#issuecomment-875238626) when the program stops running, it's a bit expensive though.

I opted to keep costs down by leveraging some recursion, iteration & branching that's already happening anyway, and folding deletion of quadtree nodes, waypoints, and colcation lists into a new function that handles gathering/printing data for the final colocation stats report.
Along the way, getting rid of some redundant iteration, counting, list construction, etc. speeds things up, eliminating `WaypointQuadtree::max_colocated()` and `Waypoint::num_colocated()` in the process, changes which also make it into the Python version.
* **C++:** The colocation report itself is down from 0.91 to 0.2 s on BiggaTomato. Add in 0.825 s for the deletions, and we're only 0.115 s above where we started. 
* **Python:** Colocation reports are faster. No deletions added to the process though; that's still all handled by the garbage collector when the program terminates. The "during" version listed in the table below has all the changes except for `WaypointQuadtree.final_report` being carved out into its own function.

  Box | Before | During | After
  | - | - | - | - |
  BT | 2.5 | 1.2 | .8
  lab1 | 1.5 | .8 | .5
  lab1.5 | 1.3 | .7 | .5
  lab2 | 1.6 | .9 | .7
  lab3 | 2.1 | 1.2 | 1.0
  lab4 | 1.9 | 1.2 | .7