@michihdeu 2021-09-21
> This and #458 are on noreaster for production site updates. Thanks.

I was confused reading "production site update" in combination to @yakra's c++ work but I realized that @yakra also modified the python code. 😌 

Will we soon switch to c++ or will we stick with using the python script for the time being?

Just as an reminder, this will have to be updated once we will switch: https://github.com/TravelMapping/DataProcessing/blob/master/SETUP.md