@yakra 2021-09-23
@michihdeu [wrote](https://github.com/TravelMapping/DataProcessing/pull/459#issuecomment-924251583):
> Will we soon switch to c++ or will we stick with the python script for the time being?

Soon, hopefully! The main thing holding us back from running the multithreaded C++ siteupdate program on noreaster is the need to [switch compilers from g++ to clang++](https://github.com/yakra/DataProcessing/issues/188). All that will need is some minor changes to the [Makefile](https://github.com/TravelMapping/DataProcessing/issues/377). I've had a new version kicking around locally for quite some time, but held off on committing it 'cause I wanted to do all my silly little benchmarks. :grin:
When I first looked into it, I was getting some results inconsistent from earlier benchmarking runs, so I put it off a while and got sidetracked by other things. Finally got the inconsistencies sorted out, and spent today *(uhh... yesterday)* arranging all the data & finishing off most of the charts (10 down, 2 or 3 to go). I'll finish those off then get the pull request in after sleeping.

Then, [a few more odds & ends before moving to production](https://github.com/TravelMapping/DataProcessing/issues/461).

After switching to clang, I want to spend some more time plugging the last of the memory leaks. This needn't *necessarily* hold us back from the C++ version moving to production, as all the memory it allocates is freed upon termination of the program, even on FreeBSD it would appear, but it's just good practice.